feat: :card_file_box: Database change from EC2 to RDS

+++ BACKEND/UserManagement/DeleteUsers/routes/user.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, HTTPException, Body
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from database import get_db
 from models import User
@@ -9,16 +9,23 @@ user_bp = APIRouter()
 @user_bp.delete("/users/{user_id}", response_model=dict)
 def delete_user(
     user_id: int, 
-    username: str = Body(..., embed=True),  # Receive the username in the request body
     db: Session = Depends(get_db),
-    is_admin: bool = Depends(validate_admin)
+    current_user: dict = Depends(validate_admin)  
 ):
-    # Search for the user by ID and name
-    user = db.query(User).filter(User.id == user_id, User.username == username).first()
+    """
+    Elimina un usuario solo si el que lo solicita es un administrador.
+    """
+    # Verificar si el usuario autenticado tiene rol de administrador
+    if current_user["role_id"] != 1:
+        raise HTTPException(status_code=403, detail="Permission denied. Only administrators can perform this action.")
+
+    # Buscar al usuario en la base de datos
+    user = db.query(User).filter(User.id == user_id).first()
     if not user:
-        raise HTTPException(status_code=404, detail="User not found or incorrect data")
-    
-    # Delete the user
+        raise HTTPException(status_code=404, detail="User not found")
+
+    # Eliminar el usuario
     db.delete(user)
     db.commit()
-    return {"message": f"User with ID {user_id} and name {username} successfully deleted"}
+
+    return {"message": f"User with ID {user_id} successfully deleted"}
